wgpu-core: Move `Device` resource methods into `device::resource`.
